R merge classes.
